+++ engine/CMakeLists.txt
@@ -13,7 +13,7 @@ file(GLOB_RECURSE headers CONFIGURE_DEPENDS "include/*.hpp")
 
 target_sources(OldEngine PUBLIC ${headers} PRIVATE ${sources})
 target_include_directories(OldEngine PUBLIC "include" "../external")
-target_link_libraries(OldEngine PUBLIC Application PUBLIC External)
+target_link_libraries(OldEngine PUBLIC Application PUBLIC External PUBLIC Physics)
 
 # ENGINE MODULES
 
@@ -21,6 +21,7 @@ message(STATUS "### Linking Engine Modules:")
 target_add_module(Engine core Core)
 target_add_module(Engine utility Utility)
 target_add_module(Engine application Application)
+target_add_module(Engine physics Physics)
 
 target_link_libraries(Engine INTERFACE OldEngine INTERFACE External)
