+++ engine/physics/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(Physics)
+module_default_init(Physics)
+
+target_link_libraries(Physics
+    PUBLIC Core
+    PUBLIC Application
+    PUBLIC OldEngine
+)
