+++ engine/source/editor.cpp
@@ -6,7 +6,7 @@
 #include "performance_tracker.hpp"
 #include "bloom_settings.hpp"
 #include "mesh.hpp"
-#include "modules/physics_module.hpp"
+#include "physics_module.hpp"
 #include "systems/physics_system.hpp"
 #include "profile_macros.hpp"
 #include "log.hpp"
